Add files via upload
Objective: This project mainly is about-face image classification. The project aims at using state of the 
art machine learning algorithms for classifying pictures of repeated faces of persons in the dataset.

Outcome: Cluster all the faces which belong to the same person. 

Techniques Performed: The dataset used here is Olivetti faces from seaborn. Performed KMeans 
Clustering to classify the faces of the people along with Random Forest Classifier to check the accuracy 
score and the number of rightly classified face.